conslidated testcase
