--- lbry/lbry/wallet/client/errors.py
@@ -1,2 +1,0 @@
-class InsufficientFundsError(Exception):
-    pass
