+++ lbry/tests/unit/blob/test_blob_file.py
@@ -2,7 +2,7 @@ import asyncio
 import tempfile
 import shutil
 import os
-from torba.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 from lbry.error import InvalidDataError, InvalidBlobHashError
 from lbry.conf import Config
 from lbry.extras.daemon.storage import SQLiteStorage
