+++ lbry/tests/unit/database/test_SQLiteStorage.py
@@ -4,7 +4,7 @@ import unittest
 import asyncio
 import logging
 import hashlib
-from torba.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 from lbry.conf import Config
 from lbry.extras.daemon.storage import SQLiteStorage
 from lbry.blob.blob_info import BlobInfo
