+++ lbry/tests/unit/dht/protocol/test_routing_table.py
@@ -1,5 +1,5 @@
 import asyncio
-from torba.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 from tests import dht_mocks
 from lbry.dht import constants
 from lbry.dht.node import Node
