+++ lbry/tests/unit/dht/test_node.py
@@ -1,7 +1,7 @@
 import asyncio
 import unittest
 import typing
-from torba.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 from tests import dht_mocks
 from lbry.conf import Config
 from lbry.dht import constants
