+++ lbry/tests/unit/stream/test_reflector.py
@@ -2,7 +2,7 @@ import os
 import asyncio
 import tempfile
 import shutil
-from torba.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 from lbry.conf import Config
 from lbry.extras.daemon.storage import SQLiteStorage
 from lbry.blob.blob_manager import BlobManager
