+++ lbry/tests/unit/wallet/test_claim_proofs.py
@@ -2,7 +2,7 @@ import unittest
 from binascii import hexlify, unhexlify
 
 from lbry.wallet.claim_proofs import get_hash_for_outpoint, verify_proof
-from lbry.wallet.client.hash import double_sha256
+from lbry.crypto.hash import double_sha256
 
 
 class ClaimProofsTestCase(unittest.TestCase):
