+++ lbry/tests/unit/wallet/test_stream_controller.py
@@ -1,6 +1,6 @@
 from lbry.wallet.stream import StreamController
 from lbry.wallet.tasks import TaskGroup
-from lbry.wallet.testcase import AsyncioTestCase
+from lbry.testcase import AsyncioTestCase
 
 
 class StreamControllerTestCase(AsyncioTestCase):
