+++ lbry/tox.ini
@@ -4,7 +4,6 @@ envlist = py37-integration
 [testenv]
 deps =
   coverage
-  ../torba
 
 extras = test
 changedir = {toxinidir}/tests
